+++ libs/nest-modules-admin/src/admin/admin.controller.ts
@@ -38,7 +38,11 @@ import {
 } from '@project-lc/nest-modules-broadcaster';
 import { GoodsService } from '@project-lc/nest-modules-goods';
 import { OrderCancelService } from '@project-lc/nest-modules-order-cancel';
-import { SellerService, SellerSettlementService } from '@project-lc/nest-modules-seller';
+import {
+  SellerService,
+  SellerSettlementInfoService,
+  SellerSettlementService,
+} from '@project-lc/nest-modules-seller';
 import {
   AdminAllLcGoodsList,
   AdminBroadcasterSettlementInfoList,
@@ -70,6 +74,7 @@ import {
   PrivacyApproachHistoryDto,
   SellerGoodsSortColumn,
   SellerGoodsSortDirection,
+  SellerSettlementTargetRes,
 } from '@project-lc/shared-types';
 import { Request } from 'express';
 import { AdminAccountService } from './admin-account.service';
@@ -86,6 +91,7 @@ export class AdminController {
     private readonly adminSettlementService: AdminSettlementService,
     private readonly adminAccountService: AdminAccountService,
     private readonly sellerSettlementService: SellerSettlementService,
+    private readonly sellerSettlementInfoService: SellerSettlementInfoService,
     private readonly orderCancelService: OrderCancelService,
     private readonly bcSettlementHistoryService: BroadcasterSettlementHistoryService,
     private readonly broadcasterSettlementInfoService: BroadcasterSettlementInfoService,
@@ -129,6 +135,14 @@ export class AdminController {
     return this.adminService.getSettlementInfo();
   }
 
+  /** 판매자 정산 등록 정보 조회 */
+  @UseGuards(JwtAuthGuard, AdminGuard)
+  @Get('/settlement/targets')
+  @Header('Cache-Control', 'no-cache, no-store, must-revalidate')
+  findSellerSettlementTargets(): Promise<SellerSettlementTargetRes> {
+    return this.sellerSettlementService.findAllSettleTargetList();
+  }
+
   /** 판매자 정산처리 */
   @UseGuards(JwtAuthGuard, AdminGuard)
   @Post('/settlement')
@@ -149,7 +163,7 @@ export class AdminController {
   /** 방송인 정산 대상 목록 조회 */
   @UseGuards(JwtAuthGuard, AdminGuard)
   @Get('/settlement/broadcaster/targets')
-  public async findSettlementTargets(
+  public async findBcSettlementTargets(
     @Query(new ValidationPipe({ transform: true })) dto: FindManyDto,
   ): Promise<BroadcasterSettlementTargets> {
     return this.settlementService.findSettlementTargets(undefined, dto);
@@ -335,7 +349,7 @@ export class AdminController {
   createSettlementConfirmHistory(
     @Body() dto: ConfirmHistoryDto,
   ): Promise<ConfirmHistory> {
-    return this.sellerSettlementService.createSettlementConfirmHistory(dto);
+    return this.sellerSettlementInfoService.createSettlementConfirmHistory(dto);
   }
 
   /** 전체 판매자 계정 목록 조회 */
