+++ libs/nest-modules-seller/src/index.ts
@@ -1,3 +1,4 @@
+export * from './seller/seller-settlement-info.service';
 export * from './seller/seller.module';
 export * from './seller/seller.service';
-export * from './seller/seller-settlement.service';
+export * from './seller/settlement/seller-settlement.service';
