+++ libs/nest-modules-seller/src/seller/seller.module.ts
@@ -4,22 +4,28 @@ import { MailVerificationModule } from '@project-lc/nest-modules-mail-verificati
 import { SellerContactsController } from './seller-contacts.controller';
 import { SellerContactsService } from './seller-contacts.service';
 import { SellerSettlementHistoryController } from './seller-settlement-history.controller';
-import { SellerSettlementService } from './seller-settlement.service';
+import { SellerSettlementInfoService } from './seller-settlement-info.service';
 import { SellerShopService } from './seller-shop.service';
 import { SellerController } from './seller.controller';
 import { SellerService } from './seller.service';
+import SellerSettlementService from './settlement/seller-settlement.service';
 
 @Module({})
 export class SellerModule {
   private static readonly providers = [
     SellerService,
     SellerSettlementService,
+    SellerSettlementInfoService,
     SellerShopService,
     SellerContactsService,
     UserPwManager,
   ];
 
-  private static readonly exports = [SellerService, SellerSettlementService];
+  private static readonly exports = [
+    SellerService,
+    SellerSettlementService,
+    SellerSettlementInfoService,
+  ];
 
   private static readonly controllers = [
     SellerController,
